refactor: renamed types
