+++ src/schema/mutation/addSubscription.mutation.ts
@@ -8,7 +8,7 @@ import mongoose from 'mongoose';
 import { Application, Channel, Form } from '../../models';
 import { AppAbility } from '../../security/defineUserAbility';
 import { createAndConsumeQueue } from '../../server/subscriberSafe';
-import { SubscriptionType } from '../types/subscription';
+import { SubscriptionType } from '../types/subscription.type';
 
 /**
  * Creates a new subscription
