+++ src/schema/mutation/editSubscription.mutation.ts
@@ -5,7 +5,7 @@ import {
   GraphQLString,
 } from 'graphql';
 import { Application } from '../../models';
-import { SubscriptionType } from '../types/subscription';
+import { SubscriptionType } from '../types/subscription.type';
 import { AppAbility } from '../../security/defineUserAbility';
 import {
   createAndConsumeQueue,
