+++ src/schema/types/apiConfiguration.type.ts
@@ -8,9 +8,9 @@ import GraphQLJSON from 'graphql-type-json';
 import { StatusEnumType, AuthEnumType } from '../../const/enumTypes';
 import { ApiConfiguration } from '../../models';
 import { AppAbility } from '../../security/defineUserAbility';
-import { AccessType } from './access';
+import { AccessType } from './access.type';
 import * as CryptoJS from 'crypto-js';
-import { Connection } from './pagination';
+import { Connection } from './pagination.type';
 import config from 'config';
 
 /** GraphQL api configuration type definition */
