+++ src/schema/types/index.ts
@@ -1,25 +1,25 @@
-export * from './access';
-export * from './apiConfiguration';
-export * from './application';
-export * from './channel';
-export * from './dashboard';
-export * from './form';
-export * from './notification';
-export * from './page';
-export * from './permission';
-export * from './positionAttribute';
-export * from './positionAttributeCategory';
-export * from './pullJob';
-export * from './record';
-export * from './referenceData';
-export * from './resource';
-export * from './role';
-export * from './setting';
-export * from './step';
-export * from './user';
-export * from './version';
-export * from './workflow';
-export * from './pagination';
-export * from './layout';
-export * from './historyVersion';
-export * from './group';
+export * from './access.type';
+export * from './apiConfiguration.type';
+export * from './application.type';
+export * from './channel.type';
+export * from './dashboard.type';
+export * from './form.type';
+export * from './notification.type';
+export * from './page.type';
+export * from './permission.type';
+export * from './positionAttribute.type';
+export * from './positionAttributeCategory.type';
+export * from './pullJob.type';
+export * from './record.type';
+export * from './referenceData.type';
+export * from './resource.type';
+export * from './role.type';
+export * from './setting.type';
+export * from './step.type';
+export * from './user.type';
+export * from './version.type';
+export * from './workflow.type';
+export * from './pagination.type';
+export * from './layout.type';
+export * from './historyVersion.type';
+export * from './group.type';
