+++ src/schema/types/page.type.ts
