+++ src/schema/types/positionAttribute.type.ts
@@ -1,6 +1,6 @@
 import { GraphQLInt, GraphQLObjectType, GraphQLString } from 'graphql';
 import { PositionAttributeCategory, User } from '../../models';
-import { PositionAttributeCategoryType } from './positionAttributeCategory';
+import { PositionAttributeCategoryType } from './positionAttributeCategory.type';
 
 /** GraphQL position attribute type definition */
 export const PositionAttributeType = new GraphQLObjectType({
