+++ src/schema/types/subscription.type.ts
@@ -1,7 +1,7 @@
 import { GraphQLObjectType, GraphQLString } from 'graphql';
 import { Channel, Form } from '../../models';
-import { ChannelType } from './channel';
-import { FormType } from './form';
+import { ChannelType } from './channel.type';
+import { FormType } from './form.type';
 import { AppAbility } from '../../security/defineUserAbility';
 
 /** GraphQL SubscriptionT type definition */
